Add new pokemon list view and show this list

+++ core/navigation/src/main/java/com/emdp/core/navigation/PokedexDestination.kt
@@ -2,5 +2,5 @@ package com.emdp.core.navigation
 
 sealed interface PokedexDestination {
     data object OpenPokemonList : PokedexDestination
-    data class OpenPokemonDetail(val id: Long) : PokedexDestination
+    data class OpenPokemonDetail(val id: Int) : PokedexDestination
 }
